# docs/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: Coeeter
    url: https://github.com/coeeter/aniways
  description: An anime tracking and streaming platform API
  title: AniWays API
  version: 1.0.0
paths:
...
  /anime/listings/genres:
    get:
      description: Get list of anime genres
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  type: string
                type: array
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get list of anime genres
      tags:
        - Anime Listings
  "/anime/listings/genres/{genre}":
    get:
      description: Get anime by genre
      parameters:
        - description: Genre name
          in: path
          name: genre
          required: true
          schema:
            type: string
        - description: Page number
          in: query
          name: page
          schema:
            type: integer
        - description: Number of items per page
          in: query
          name: itemsPerPage
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.AnimeListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get anime by genre
      tags:
        - Anime Listings
  /anime/listings/popular:
    get:
      description: Get popular anime
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/models.AnimeResponse"
                type: array
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get popular anime
      tags:
        - Anime Listings
  /anime/listings/random:
    get:
      description: Get random anime
      parameters:
        - description: Optional genre filter
          in: query
          name: genre
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.AnimeResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get random anime
      tags:
        - Anime Listings
  /anime/listings/recently-updated:
    get:
      description: Get recently updated anime
      parameters:
        - description: Page number
          in: query
          name: page
          schema:
            type: integer
        - description: Number of items per page
          in: query
          name: itemsPerPage
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.AnimeListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get recently updated anime
      tags:
        - Anime Listings
  /anime/listings/search:
    get:
      description: Search anime
      parameters:
        - description: Search query (minimum 3 characters)
          in: query
          name: q
          required: true
          schema:
            type: string
        - description: Optional genre filter (minimum 3 characters)
          in: query
          name: genre
          schema:
            type: string
        - description: Page number
          in: query
          name: page
          schema:
            type: integer
        - description: Number of items per page
          in: query
          name: itemsPerPage
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.AnimeListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Search anime
      tags:
        - Anime Listings
  /anime/listings/seasonal:
    get:
      description: Get seasonal anime
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/models.SeasonalAnimeResponse"
                type: array
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get seasonal anime
      tags:
        - Anime Listings
  /anime/listings/seasons:
    get:
      description: Get anime by season and/or year
      parameters:
        - description: Season filter
          in: query
          name: season
          schema:
            type: string
        - description: Year filter
          in: query
          name: year
          schema:
            type: integer
        - description: Page number
          in: query
          name: page
          schema:
            type: integer
        - description: Number of items per page
          in: query
          name: itemsPerPage
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.AnimeListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get anime by season and/or year
      tags:
        - Anime Listings
  /anime/listings/trending:
    get:
      description: Get trending anime
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/models.AnimeResponse"
                type: array
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get trending anime
      tags:
        - Anime Listings
...
  "/auth/oauth/{provider}":
    get:
      description: Start OAuth authentication
      parameters:
        - description: OAuth provider
          in: path
          name: provider
          required: true
          schema:
            type: string
        - description: Redirect URL after authentication
          in: query
          name: redirect
          schema:
            type: string
      responses:
        "302":
          description: Found
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      security:
        - cookieAuth: []
      summary: Start OAuth authentication
      tags:
        - OAuth
  "/auth/oauth/{provider}/callback":
    get:
      description: OAuth callback endpoint
      parameters:
        - description: OAuth provider
          in: path
          name: provider
          required: true
          schema:
            type: string
        - description: OAuth authorization code
          in: query
          name: code
          required: true
          schema:
            type: string
        - description: OAuth state parameter
          in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Found
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      security:
        - cookieAuth: []
      summary: OAuth callback endpoint
      tags:
        - OAuth
...
servers:
  - url: http://localhost:8080
components:
...
  securitySchemes:
    cookieAuth:
      in: cookie
      name: aniways_session
      type: apiKey
  schemas:
    models.AnimeListResponse:
      properties:
        items:
          items:
            $ref: "#/components/schemas/models.AnimeResponse"
          type: array
        pageInfo:
          $ref: "#/components/schemas/models.PageInfo"
      type: object
...
    models.AnimeResponse:
      properties:
        anilistId:
          example: 67890
          type: integer
        ename:
          example: Attack on Titan
          type: string
        genre:
          example: Action, Drama
          type: string
        id:
          example: V1StGXR8Z5jdHi6B
          type: string
        imageUrl:
          example: https://example.com/anime/image.jpg
          type: string
        jname:
          example: 進撃の巨人
          type: string
        lastEpisode:
          example: 25
          type: integer
        malId:
          example: 12345
          type: integer
        season:
          example: spring
          type: string
        seasonYear:
          example: 2023
          type: integer
      type: object
...
    models.ErrorResponse:
      properties:
        error:
          example: Invalid request
          type: string
      type: object
...
    models.PageInfo:
      properties:
        currentPage:
          type: integer
        hasNextPage:
          type: boolean
        hasPrevPage:
          type: boolean
        totalPages:
          type: integer
      type: object
...
    models.SeasonalAnimeResponse:
      properties:
        anime:
          $ref: "#/components/schemas/models.AnimeResponse"
        bannerImageUrl:
          example: https://example.com/banner.jpg
          type: string
        description:
          example: New seasonal anime description
          type: string
        episodes:
          example: 12
          type: integer
        id:
          example: V1StGXR8Z5jdHi6B
          type: string
        startDate:
          example: 1672531200
          type: integer
        type:
          example: TV
          type: string
      type: object
...